requirements.txt
kfp
xgboost
pandas
numpy
scikit-learn
mlflow
matplotlib
pyyaml
kubernetes
